feat: Add custom index.html
This commit introduces a custom `index.html` file for the web build,
replacing the default Emscripten-generated page. The `web.sh` script
has been updated to copy the new file and output a JavaScript file
instead of an HTML file. This provides more control over the user
interface and presentation of the simulation.